% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_read}
\alias{data_read}
\title{上传数据}
\usage{
data_read(
  conn = tsda::conn_rds("cprds"),
  file_name = "data-raw/data/产品大类模板.xlsx",
  sheet_name = "产品大类",
  table_name = "rds_mtrl_prdCategory",
  table_key = "FPrdCategoryNumber",
  table_caption = "产品大类代码",
  pageCount = 500
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}

\item{sheet_name}{页签名}

\item{table_name}{表名}

\item{table_key}{表主键}

\item{table_caption}{表标题}

\item{pageCount}{分页数}
}
\value{
返回值
}
\description{
上传数据
}
\examples{
data_read()
}
