% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_read2}
\alias{data_read2}
\title{上传数据}
\usage{
data_read2(
  conn = tsda::conn_rds("cprds"),
  file_name = "data-raw/data/属性类别配置模板.xlsx",
  sheet_name = "属性类别配置",
  table_name = "rds_mtrl_propCategoryConfig",
  table_key = "FPrdCategoryNumber",
  table_key2 = "FPropCategoryNumber",
  table_caption = "产品大类代码",
  table_caption2 = "物料属性类别代码",
  pageCount = 500
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}

\item{sheet_name}{页签名}

\item{table_name}{表名}

\item{table_key}{表主键}

\item{table_key2}{主键2}

\item{table_caption}{表标题}

\item{table_caption2}{标题2}

\item{pageCount}{分页数}
}
\value{
返回值
}
\description{
上传数据
}
\examples{
data_read()
}
