% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_read3}
\alias{data_read3}
\title{上传数据}
\usage{
data_read3(
  conn_cfg = "config/conn_cfg.R",
  file_name = "data-raw/data/属性选项配置模板.xlsx",
  sheet_name = "属性选项配置",
  table_name = "rds_mtrl_propValueConfig",
  table_key = "FPrdCategoryNumber",
  table_key2 = "FPropCategoryNumber",
  table_key3 = "FPropNumber",
  table_caption = "产品大类代码",
  table_caption2 = "属性类别代码",
  table_caption3 = "属性选项代码",
  pageCount = 500
)
}
\arguments{
\item{conn_cfg}{连接}

\item{sheet_name}{页签名}

\item{table_name}{表名}

\item{table_key}{表主键}

\item{table_key2}{主键2}

\item{table_key3}{键3}

\item{table_caption}{表标题}

\item{table_caption2}{标题2}

\item{table_caption3}{标题3}

\item{pageCount}{分页数}
}
\value{
返回值
}
\description{
上传数据
}
\examples{
data_read()
}
