% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_upload}
\alias{data_upload}
\title{上传元数据}
\usage{
data_upload(
  conn_cfg = "config/conn_cfg.R",
  data,
  FTableName = "rds_mtrl_prdConfigRes"
)
}
\arguments{
\item{conn_cfg}{连接文件}

\item{data}{数据}

\item{FTableName}{表名}
}
\value{
返回值
}
\description{
上传元数据
}
\examples{
meta_upload()
}
