% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_createTable}
\alias{meta_createTable}
\title{创建表格}
\usage{
meta_createTable(conn = tsda::conn_rds("metadata"), FTableCaption = "产品大类")
}
\arguments{
\item{conn}{连接}

\item{FTableCaption}{表名}
}
\value{
返回值
}
\description{
创建表格
}
\examples{
meta_createTable()
}
