% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_createView}
\alias{meta_createView}
\title{创建视图}
\usage{
meta_createView(conn = tsda::conn_rds("metadata"), FTableCaption = "产品大类")
}
\arguments{
\item{conn}{连接}

\item{FTableCaption}{表名}
}
\value{
返回值
}
\description{
创建视图
}
\examples{
meta_createView()
}
