% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_maxId}
\alias{meta_maxId}
\title{查询表的最大值}
\usage{
meta_maxId(
  conn = tsda::conn_rds("metadata"),
  FTableName = "t_dict_multipleMaterial"
)
}
\arguments{
\item{conn}{连接}

\item{FTableName}{表名}
}
\value{
返回值
}
\description{
查询表的最大值
}
\examples{
meta_maxId()
}
