% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdCategory.R
\name{prdCategory_read}
\alias{prdCategory_read}
\title{读取产品大类数据}
\usage{
prdCategory_read(
  conn = tsda::conn_rds("cprds"),
  file_name = "data-raw/data/rds_mtrl_productCategory.xlsx",
  sheet_name = "产品大类"
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}

\item{sheet_name}{页签表}
}
\value{
返回值
}
\description{
读取产品大类数据
}
\examples{
prdCategory_read()
}
