% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdCategory.R
\name{prdCategoryServer_query}
\alias{prdCategoryServer_query}
\title{产品大类查询}
\usage{
prdCategoryServer_query(input, output, session, conn_cfg)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn_cfg}{连接配置}
}
\value{
返回值
}
\description{
产品大类查询
}
\examples{
prdCategoryServer_query()
}
