% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdCategory.R
\name{prdCategoryUIServer_sheet}
\alias{prdCategoryUIServer_sheet}
\title{文本选择}
\usage{
prdCategoryUIServer_sheet(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
文本选择
}
\examples{
prdCategoryUIServer_sheet()
}
