% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdConfigRes.R
\name{prdConfigRes_isNew}
\alias{prdConfigRes_isNew}
\title{判断是否是最新}
\usage{
prdConfigRes_isNew(
  conn_cfg = "config/conn_cfg.R",
  FPrdGroupNumber = "1.1.1.004.002",
 
    FPropValueNumberFlag = "C000001_AC0001_AD0003_AF0001_AH0002_AJ0011_AM0001_AS0001_AT0001"
)
}
\arguments{
\item{conn_cfg}{连接配置}

\item{FPrdGroupNumber}{产品分组代码}

\item{FPropValueNumberFlag}{选配结果代码}
}
\value{
返回值
}
\description{
判断是否是最新
}
\examples{
prdConfigRes_isNew()
}
