% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGen.R
\name{prdGen_prdGroup_reactiveData}
\alias{prdGen_prdGroup_reactiveData}
\title{生成动态数据}
\usage{
prdGen_prdGroup_reactiveData(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{返回值}
}
\value{
返回值
}
\description{
生成动态数据
}
\examples{
prdGen_prdGroup_reactiveData()
}
