% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenList.R
\name{prdGenList_isNew}
\alias{prdGenList_isNew}
\title{判断是新的数据}
\usage{
prdGenList_isNew(
  conn = tsda::conn_rds("cprds"),
  FPrdGroupNumber = "1.1.1.004.002",
 
    FPropValueNumberFlag = "C000001_AC0001_AD0003_AF0001_AH0002_AJ0011_AM0001_AS0001_AT0001"
)
}
\arguments{
\item{conn}{连接}

\item{FPrdGroupNumber}{产品分组代码}

\item{FPropValueNumberFlag}{产品属性标签}
}
\value{
返回值
}
\description{
判断是新的数据
}
\examples{
prdGenList_isNew()
}
