% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenList.R
\name{prdGenList_maxId}
\alias{prdGenList_maxId}
\title{获取最大号ID}
\usage{
prdGenList_maxId(conn = tsda::conn_rds("cprds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取最大号ID
}
\examples{
prdGenList_maxId()
}
