% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenList.R
\name{prdGenList_query}
\alias{prdGenList_query}
\title{产品结果查看}
\usage{
prdGenList_query(
  conn_cfg = "config/conn_cfg.R",
  FPrdGroupNumber = "1.1.1.004.002",
 
    FPropValueNumberFlag = "C000001_AC0001_AD0003_AF0001_AH0002_AJ0011_AM0001_AS0001_AT0001"
)
}
\arguments{
\item{conn_cfg}{连接}

\item{FPrdGroupNumber}{产品分组}

\item{FPropValueNumberFlag}{选配结果代码}
}
\value{
返回值
}
\description{
产品结果查看
}
\examples{
prdGenList_query()
}
