% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenList.R
\name{prdGenList_upload}
\alias{prdGenList_upload}
\title{产品结果查看}
\usage{
prdGenList_upload(
  conn_cfg = "config/conn_cfg.R",
  FPrdGroupNumber = "1.1.1.004.002",
  FPrdGroupName = "PCR板无群边",
 
    FPropValueNumberFlag = "C000001_AC0001_AD0003_AF0001_AH0002_AJ0011_AM0001_AS0001_AT0001",
 
    FPropValueNameFlag = "赛普_透明色_耐高温_袋装_40ul_96孔_单色注塑_R_印刷"
)
}
\arguments{
\item{conn_cfg}{连接设置}

\item{FPrdGroupNumber}{产品分组}

\item{FPrdGroupName}{产品分组名称}

\item{FPropValueNumberFlag}{选配结果代码}

\item{FPropValueNameFlag}{选配结果名称}
}
\value{
返回值
}
\description{
产品结果查看
}
\examples{
prdGenList_upload()
}
