% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenRule.R
\name{prdGenRule_autoNumber}
\alias{prdGenRule_autoNumber}
\title{获取下一个代码}
\usage{
prdGenRule_autoNumber(
  conn = tsda::conn_rds("cprds"),
  FPrdGroupNumber = "1.1.1.001.001",
  FNumberLength = 6
)
}
\arguments{
\item{conn}{连接}

\item{FPrdGroupNumber}{物料分组代码}

\item{FNumberLength}{长度}
}
\value{
返回值
}
\description{
获取下一个代码
}
\examples{
prdGenRule_autoNumber()
}
