% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenRule.R
\name{prdGenRule_isNew}
\alias{prdGenRule_isNew}
\title{是否最新}
\usage{
prdGenRule_isNew(
  conn = tsda::conn_rds("cprds"),
  FPrdGroupNumber = "1.1.1.001.001"
)
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
是否最新
}
\examples{
prdGenRule_isNew()
}
