% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenRule.R
\name{prdGenRule_maxId}
\alias{prdGenRule_maxId}
\title{获取最大号ID}
\usage{
prdGenRule_maxId(conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{conn_cfg}{连接配置}
}
\value{
返回值
}
\description{
获取最大号ID
}
\examples{
prdGenRule_maxId()
}
