% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenRule.R
\name{prdGenRule_SetNextNumber}
\alias{prdGenRule_SetNextNumber}
\title{设置下一个代码}
\usage{
prdGenRule_SetNextNumber(
  conn = tsda::conn_rds("cprds"),
  FPrdGroupNumber = "1.1.1.001.001",
  FNextNumber = 1
)
}
\arguments{
\item{conn}{连接}

\item{FPrdGroupNumber}{物料分组代码}

\item{FNextNumber}{设置下一个值}
}
\value{
返回值
}
\description{
设置下一个代码
}
\examples{

prdGenRule_SetNextNumber()
}
