% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGen.R
\name{prdGenServer}
\alias{prdGenServer}
\title{处理逻辑}
\usage{
prdGenServer(input, output, session, conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn_cfg}{连接文件}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
materialGeneratorServer()
}
