% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGen.R
\name{prdGenServer_prdCategory}
\alias{prdGenServer_prdCategory}
\title{处理逻辑}
\usage{
prdGenServer_prdCategory(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
materialGeneratorServer()
}
