% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroup_query}
\alias{prdGroup_query}
\title{产品大类查询}
\usage{
prdGroup_query(conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{conn_cfg}{连接设置}
}
\value{
返回值
}
\description{
产品大类查询
}
\examples{
prdGroup_query()
}
