% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroup_queryDetail}
\alias{prdGroup_queryDetail}
\title{产品大类查询}
\usage{
prdGroup_queryDetail(
  conn = tsda::conn_rds("cprds"),
  FPrdCategoryFNumber = "1.1.1.006"
)
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
产品大类查询
}
\examples{
prdGroup_query()
}
