% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroupServer}
\alias{prdGroupServer}
\title{产品大娄的处理函数}
\usage{
prdGroupServer(input, output, session, conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn}{连接}
}
\value{
返回值
}
\description{
产品大娄的处理函数
}
\examples{
prdGroupServer()
}
