% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroupServer_query}
\alias{prdGroupServer_query}
\title{产品大类查询}
\usage{
prdGroupServer_query(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
产品大类查询
}
\examples{
prdGroupServer_query()
}
