% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroupUIServer_sheet}
\alias{prdGroupUIServer_sheet}
\title{文本选择}
\usage{
prdGroupUIServer_sheet(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
文本选择
}
\examples{
prdGroupUIServer_sheet()
}
