% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdSel.R
\name{prdSel_query}
\alias{prdSel_query}
\title{属性类别查询}
\usage{
prdSel_query(conn = tsda::conn_rds("cprds"), FNumber = "")
}
\arguments{
\item{conn}{连接}

\item{FNumber}{}
}
\value{
返回值
}
\description{
属性类别查询
}
\examples{
propCategory_query()
}
