% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdSel.R
\name{prdSelServer}
\alias{prdSelServer}
\title{处理逻辑}
\usage{
prdSelServer(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
prdSelServer()
}
