% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategory.R
\name{propCategory_query}
\alias{propCategory_query}
\title{属性类别查询}
\usage{
propCategory_query(conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{conn_cfg}{连接配置}
}
\value{
返回值
}
\description{
属性类别查询
}
\examples{
propCategory_query()
}
