% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategory.R
\name{propCategory_queryDetail}
\alias{propCategory_queryDetail}
\title{属性类别查询}
\usage{
propCategory_queryDetail(
  conn_cfg = "config/conn_cfg.R",
  FPrdCategoryNumber = "1.1.1.001"
)
}
\arguments{
\item{conn_cfg}{连接}

\item{FPrdCategoryNumber}{代码}
}
\value{
返回值
}
\description{
属性类别查询
}
\examples{
propCategory_queryDetail()
}
