% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategory.R
\name{propCategory_read}
\alias{propCategory_read}
\title{读取属性类别数据}
\usage{
propCategory_read(
  conn = tsda::conn_rds("cprds"),
  file_name = "data-raw/data/属性类别模板.xlsx",
  sheet_name = "属性类别"
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}

\item{sheet_name}{页签表}
}
\value{
返回值
}
\description{
读取属性类别数据
}
\examples{
propCategory_read()
}
