% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategoryConfig.R
\name{propCategoryConfig_query}
\alias{propCategoryConfig_query}
\title{属性类别配置查询}
\usage{
propCategoryConfig_query(conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{conn_cfg}{连接配置}
}
\value{
返回值
}
\description{
属性类别配置查询
}
\examples{
propCategoryConfig_query()
}
