% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategoryConfig.R
\name{propCategoryConfig_read}
\alias{propCategoryConfig_read}
\title{读取属性类别数据}
\usage{
propCategoryConfig_read(
  conn_cfg = "config/conn_cfg.R",
  file_name = "data-raw/data/属性类别配置模板.xlsx",
  sheet_name = "属性类别配置"
)
}
\arguments{
\item{conn_cfg}{连接设置}

\item{file_name}{文件名}

\item{sheet_name}{页签表}
}
\value{
返回值
}
\description{
读取属性类别数据
}
\examples{
propCategoryConfig_read()
}
