% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategoryConfig.R
\name{propCategoryConfigServer}
\alias{propCategoryConfigServer}
\title{产品大娄的处理函数}
\usage{
propCategoryConfigServer(input, output, session, conn)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn}{连接}
}
\value{
返回值
}
\description{
产品大娄的处理函数
}
\examples{
propCategoryConfigServer()
}
