% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategoryConfig.R
\name{propCategoryConfigServer_query}
\alias{propCategoryConfigServer_query}
\title{属性类别配置查询}
\usage{
propCategoryConfigServer_query(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
属性类别配置查询
}
\examples{
propCategoryConfigServer_query()
}
