% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategory.R
\name{propCategoryServer_query}
\alias{propCategoryServer_query}
\title{属性类别查询}
\usage{
propCategoryServer_query(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
属性类别查询
}
\examples{
propCategoryServer_query()
}
