% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValue.R
\name{propValue_query}
\alias{propValue_query}
\title{属性选项查询}
\usage{
propValue_query(conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{conn_cfg}{连接设置}
}
\value{
返回值
}
\description{
属性选项查询
}
\examples{
propValue_query()
}
