% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValue.R
\name{propValue_queryByNumber}
\alias{propValue_queryByNumber}
\title{返回选项值}
\usage{
propValue_queryByNumber(
  conn_cfg = "config/conn_cfg.R",
  FPropNumber = "C000009"
)
}
\arguments{
\item{conn_cfg}{连接设置}

\item{FPropNumber}{选项代码}
}
\value{
返回值
}
\description{
返回选项值
}
\examples{
propValue_queryByNumber()
}
