% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValue.R
\name{propValue_queryDetail}
\alias{propValue_queryDetail}
\title{属性选项查询}
\usage{
propValue_queryDetail(
  conn_cfg = "config/conn_cfg.R",
  FPrdCategoryNumber = "1.1.1.001",
  FPropCategoryNumber = "AF"
)
}
\arguments{
\item{conn_cfg}{连接设置}

\item{FPrdCategoryNumber}{产品大类代码}

\item{FPropCategoryNumber}{产品属性代码}
}
\value{
返回值
}
\description{
属性选项查询
}
\examples{
propValue_queryDetail()
}
