% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValueConfig.R
\name{propValueConfig_query}
\alias{propValueConfig_query}
\title{属性选项配置查询}
\usage{
propValueConfig_query(conn = tsda::conn_rds("cprds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
属性选项配置查询
}
\examples{
propValueConfig_query()
}
