% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValueConfig.R
\name{propValueConfigServer}
\alias{propValueConfigServer}
\title{产品大娄的处理函数}
\usage{
propValueConfigServer(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
产品大娄的处理函数
}
\examples{
propValueConfigServer()
}
