% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValueConfig.R
\name{propValueConfigServer_query}
\alias{propValueConfigServer_query}
\title{属性选项配置查询}
\usage{
propValueConfigServer_query(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
属性选项配置查询
}
\examples{
propValueConfigServer_query()
}
