% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValueConfig.R
\name{propValueConfigUIServer_sheet}
\alias{propValueConfigUIServer_sheet}
\title{文本选择}
\usage{
propValueConfigUIServer_sheet(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
文本选择
}
\examples{
propValueConfigUIServer_sheet()
}
