% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValue.R
\name{propValueServer}
\alias{propValueServer}
\title{产品大娄的处理函数}
\usage{
propValueServer(input, output, session, conn_cfg)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn_cfg}{连接设置}
}
\value{
返回值
}
\description{
产品大娄的处理函数
}
\examples{
propValueServer()
}
