% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popValue.R
\name{propValueServer_query}
\alias{propValueServer_query}
\title{属性选项查询}
\usage{
propValueServer_query(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
属性选项查询
}
\examples{
propValueServer_query()
}
