% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popValue.R
\name{propValueUIServer_sheet}
\alias{propValueUIServer_sheet}
\title{文本选择}
\usage{
propValueUIServer_sheet(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
文本选择
}
\examples{
propValueUIServer_sheet()
}
