% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabPanel_initial_Server}
\alias{tabPanel_initial_Server}
\title{动态标签的加载}
\usage{
tabPanel_initial_Server(
  input,
  output,
  session,
  conn_cfg,
  app_id = "cpdms",
  role_name = "admin"
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn_cfg}{会话连接}

\item{app_id}{会话IP}

\item{role_name}{角色}
}
\value{
返回值
}
\description{
动态标签的加载
}
\examples{
tabPanel_initial_Server()
}
