% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdCategory.R
\name{prdCategory_query}
\alias{prdCategory_query}
\title{产品大类查询}
\usage{
prdCategory_query(conn = tsda::conn_rds("cprds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
产品大类查询
}
\examples{
prdCategory_query()
}
