% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenList.R
\name{prdGenList_maxId}
\alias{prdGenList_maxId}
\title{获取最大号ID}
\usage{
prdGenList_maxId(conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{conn_cfg}{连接文件}
}
\value{
返回值
}
\description{
获取最大号ID
}
\examples{
prdGenList_maxId()
}
