% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenRule.R
\name{prdGenRule_getNextNumber}
\alias{prdGenRule_getNextNumber}
\title{获取下一个代码}
\usage{
prdGenRule_getNextNumber(
  conn_cfg = "config/conn_cfg.R",
  FPrdGroupNumber = "1.1.1.001.001",
  FNumberLength = 6
)
}
\arguments{
\item{conn_cfg}{连接文件}

\item{FPrdGroupNumber}{物料分组代码}

\item{FNumberLength}{长度}
}
\value{
返回值
}
\description{
获取下一个代码
}
\examples{
prdGenRule_getNextNumber()
}
