% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGenRule.R
\name{prdGenRule_isNew}
\alias{prdGenRule_isNew}
\title{是否最新}
\usage{
prdGenRule_isNew(
  conn_cfg = "config/conn_cfg.R",
  FPrdGroupNumber = "1.1.1.001.001"
)
}
\arguments{
\item{conn_cfg}{连接配置}

\item{FPrdGroupNumber}{分组代码}
}
\value{
返回值
}
\description{
是否最新
}
\examples{
prdGenRule_isNew()
}
