% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroup_read}
\alias{prdGroup_read}
\title{读取产品分组数据}
\usage{
prdGroup_read(
  conn_cfg = "config/conn_cfg.R",
  file_name = "data-raw/data/产品分组模板.xlsx",
  sheet_name = "产品大类"
)
}
\arguments{
\item{conn_cfg}{连接设置}

\item{file_name}{文件名}

\item{sheet_name}{页签表}
}
\value{
返回值
}
\description{
读取产品分组数据
}
\examples{
prdGroup_read()
}
