% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroupServer_read}
\alias{prdGroupServer_read}
\title{处理内容}
\usage{
prdGroupServer_read(input, output, session, conn_cfg = "config/conn_cfg.R")
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会议}

\item{conn_cfg}{连接设置}
}
\value{
返回值
}
\description{
处理内容
}
\examples{
file_deal()
}
