% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdSel.R
\name{prdSel_query}
\alias{prdSel_query}
\title{属性类别查询}
\usage{
prdSel_query(conn_cfg = "config/conn_cfg.R", FNumber = "")
}
\arguments{
\item{conn_cfg}{连接文件}

\item{FNumber}{代码}
}
\value{
返回值
}
\description{
属性类别查询
}
\examples{
propCategory_query()
}
