% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategory.R
\name{propCategory_queryDetail}
\alias{propCategory_queryDetail}
\title{属性类别查询}
\usage{
propCategory_queryDetail(
  conn = tsda::conn_rds("cprds"),
  FPrdCategoryNumber = "1.1.1.001"
)
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
属性类别查询
}
\examples{
propCategory_queryDetail()
}
