% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValueConfig.R
\name{propValueConfig_read}
\alias{propValueConfig_read}
\title{读取属性选项数据}
\usage{
propValueConfig_read(
  conn = tsda::conn_rds("cprds"),
  file_name = "data-raw/data/属性选项配置模板.xlsx",
  sheet_name = "属性选项配置"
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}

\item{sheet_name}{页签表}
}
\value{
返回值
}
\description{
读取属性选项数据
}
\examples{
propValueConfig_read()
}
